Move commonly used stuff to __init__.py
